DEV: sig verification updates

+++ contracts/EnigmaSimulation.sol
@@ -192,7 +192,7 @@ contract EnigmaSimulation is EnigmaStorage, EnigmaEvents, Getters {
     */
     function deploySecretContractFailure(
         bytes32 _taskId,
-        uint _gasUsed,
+        uint64 _gasUsed,
         bytes memory _sig
     )
     public
@@ -221,7 +221,7 @@ contract EnigmaSimulation is EnigmaStorage, EnigmaEvents, Getters {
         bytes32 _initStateDeltaHash,
         bytes memory _optionalEthereumData,
         address _optionalEthereumContractAddress,
-        uint _gasUsed,
+        uint64 _gasUsed,
         bytes memory _sig
     )
     public
@@ -357,7 +357,7 @@ contract EnigmaSimulation is EnigmaStorage, EnigmaEvents, Getters {
         bytes32 _outputHash,
         bytes memory _optionalEthereumData,
         address _optionalEthereumContractAddress,
-        uint _gasUsed,
+        uint64 _gasUsed,
         bytes memory _sig
     )
     public
@@ -388,7 +388,7 @@ contract EnigmaSimulation is EnigmaStorage, EnigmaEvents, Getters {
         bytes32[] memory _outputHashes,
         bytes memory _optionalEthereumData,
         address _optionalEthereumContractAddress,
-        uint[] memory _gasesUsed,
+        uint64[] memory _gasesUsed,
         bytes memory _sig
     )
     public
@@ -411,7 +411,7 @@ contract EnigmaSimulation is EnigmaStorage, EnigmaEvents, Getters {
     function commitTaskFailure(
         bytes32 _scAddr,
         bytes32 _taskId,
-        uint _gasUsed,
+        uint64 _gasUsed,
         bytes memory _sig
     )
     public
